Move form field to own file
